test float32

+++ extra_models/eff_hrnet.py
@@ -116,7 +116,9 @@ class EfficientHRNet_B0(EfficientHRNet):
         effnet=keras.applications.EfficientNetB0(
             include_top=False,
             weights=None,
-            input_shape=input_shape[1:]
+            input_shape=input_shape[1:],
+            #TODO: Test dtype
+            dtype=tf.float32,
         )
         self.backbone = keras.Model(
             inputs = effnet.input,
